Accept the async handler type as-is.
Right now, the Handler typed is forced into a pointer. But I think the Handler
type should be used as-is, leaving control to the caller. The case for a
non-pointer type is when the Handler is merely a wrapper. I want to do:

```zig
const TLSHandler struct {
  io: *IO,

  fn send(self: TLSHandler, data: []const u8) !void {
    return self.io.send(data);
  }

  //...
}
```

In which case, forcing a pointer isn't helpful.