don't change handshake to application cipher on short read
While parsing client flight 2 on server it first calculates application
cipher then parses received client flight 2 with existing handshake
cipher. We need to change cipher even in the case of the error because
it is used later to create alert from that error.

But in the case of short read changing cipher from handshake to
application will break parsing client flight on the next read.

Fixes #12